Added a choice for the user to source setup.bash
Added "environment_setup_bashrc" function, which let's user choose whether sourcing setup.bash file should be added to bashrc.
In the Readme file of ros installation, it says:
"After this it will ask you to source your .bashrc. This is needed so all ROS commands can be used from anywhere in the terminal and for auto-completion of commands. A backup will be created in the file .bashrc_before_ros."

But it doesn't do that. So I added it following the same style exactly.